// src/app/shared/models/assignment.model.ts
export interface Assignment {
  userName?: string;
  work02assign0101?: boolean;
	work02assign0102?: boolean;
	work02assign0103?: boolean;
	work02assign0104?: boolean;
	work02assign0105?: boolean;
	work02assign0106?: boolean;
	work02assign0107?: boolean;
	work02assign0108?: boolean;
	work02assign0109?: boolean;
	work02assign0110?: boolean;
	work02assign0201?: boolean;
	work02assign0202?: boolean;
	work02assign0203?: boolean;
	work02assign0204?: boolean;
	work02assign0205?: boolean;
	work02assign0206?: boolean;
	work02assign0207?: boolean;
	work02assign0208?: boolean;
	work02assign0301?: boolean;
	work02assign0302?: boolean;
	work02assign0303?: boolean;
	work02assign0304?: boolean;
	work02assign0305?: boolean;
	work02assign0401?: boolean;
	work02assign0402?: boolean;
	work02assign0403?: boolean;
	work02assign0404?: boolean;
	work02assign0405?: boolean;
	work02assign0406?: boolean;
	work02assign0407?: boolean;
	work02assign0501?: boolean;
	work02assign0502?: boolean;
	work02assign0503?: boolean;
	work02assign0504?: boolean;
	work02assign0505?: boolean;
	work02assign0601?: boolean;
	work02assign0602?: boolean;
	work02assign0603?: boolean;
	work02assign0604?: boolean;
	work02assign0701?: boolean;
	work02assign0702?: boolean;
	work02assign0703?: boolean;
}